Update 09-14-2022
Continue work on hoxel control. Found that only the Z command was being executed and that the current hardware configuration will not allow for continuous and simultaneous multi-DoF control of the device

Found that the hoxels' dynamic response is so good that the serial commands are too noisy. They cause too much vibrotactile stimulation. Need to work on adding some sort of filter to the signal. Maybe IIR?

Readjusted the GUI so that the serial commands are output to the GUI. Aspect ratio needs to be made adjustable.